+++ package.json
@@ -23,6 +23,7 @@
     "babel-preset-env": "^1.7.0",
     "babel-register": "^6.26.0",
     "blue-tape": "^1.0.0",
+    "fast-deep-equal": "^2.0.1",
     "stats-webpack-plugin": "^0.4.3",
     "webpack": "^4.16.1",
     "webpack-cli": "^3.1.0",
